backend/requirements.txt
fastapi>=0.95.0,<0.100.0
uvicorn>=0.22.0,<0.23.0
pydantic>=1.10.0,<2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0,<2.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=4.0.0